Plot data in the isotope space

This function represents the data in the isotope space to overlay consumer and sources isotope values. User can also use the MixSIAR::plot_data().

plot_isospace(
  mix,
  source,
  discr,
  filename = "isospace",
  plot_save_pdf = FALSE,
  plot_save_png = FALSE
)

Arguments

mix
a named list. The output returned by the MixSIAR::load_mix_data() function and containing consumer isotope values.

source
a named list. The output returned by the MixSIAR::load_source_data() function and containing mean and standard deviation isotope values of sources (and in some case raw values).

discr
a named list. The output returned by the MixSIAR::load_discr_data() function and containing discrimination values.

filename
a character of length 1. The name of the plot file to save.

plot_save_pdf
a logical value. If TRUE, the plot is saved as pdf.

plot_save_png
a logical value. If TRUE, the plot is saved as png.

See also

Examples

if (FALSE) {
## Please visit the vignette available at: 
## https://ahasverus.github.io/samplesim/articles/samplesim.html
}